File

+++ ChatApplication/UserControls/FileSenderPage.cs
@@ -35,7 +35,10 @@ namespace ChatApplication.UserControls
             {
                 string NetworkPath = @"\\SPARE-B11\Chat Application Profile\";
                 string newfilePath = Path.Combine(NetworkPath, Path.GetFileNameWithoutExtension(path) + Path.GetExtension(path));
-                File.Copy(path, newfilePath, true);
+                if (!File.Exists(newfilePath))
+                {
+                    File.Copy(path, newfilePath, true); 
+                }
                 FileMsgReady?.Invoke(this, path);
             }
             Hide();
